Split 'UserData' into 'Create', 'Read', and 'Update', and removed 'ReadById' since it isn't needed.
